pyproject.toml
[project]
name = "fastapi-agent"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "aisqlite>=1.1.0",
    "alembic>=1.18.4",
    "ddgs>=9.11.3",
    "fastapi>=0.135.1",
    "greenlet>=3.3.2",
    "openai-agents[litellm]>=0.10.5",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.5",
    "scalar-fastapi>=1.8.0",
    "sqlmodel>=0.0.37",
    "uvicorn>=0.41.0",
]
[dependency-groups]
dev = [
    "ruff>=0.15.5",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
exclude = ["alembic"]
